requirements.txt
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==20.9
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dotenv==0.18.0
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.18
urllib3==1.26.5
webencodings==0.5.1
Werkzeug==2.0.1
